CON-450: remove ProtoUtil.creatorJustification(block: Block) and move EquivocationDetector to 'equivocations' package
